+++ src/components/common-components/input/GrayInput.tsx
