+++ src/components/common-components/input/GrayInput.types.ts
